cloud-functions/batch-processor-function/requirements.txt
functions-framework>=3.0.0
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
pandas>=1.5.0
requests>=2.25.0
db-dtypes>=1.0.0